+++ lib/apis/turbo_firestore_update_api.dart
@@ -63,7 +63,7 @@ extension TurboFirestoreUpdateApi<T> on TurboFirestoreApi<T> {
   /// ```
   ///
   /// See also:
-  /// [updateDocs] batch updates
+  /// [updateDocInBatch] batch updates
   /// [createDoc] document creation
   Future<TurboResponse<DocumentReference>> updateDoc({
     required TurboWriteable writeable,
@@ -84,8 +84,7 @@ extension TurboFirestoreUpdateApi<T> on TurboFirestoreApi<T> {
         message: 'Checking if writeable is valid..',
         sensitiveData: null,
       );
-      final TurboResponse<DocumentReference>? invalidResponse =
-          writeable.validate();
+      final TurboResponse<DocumentReference>? invalidResponse = writeable.validate();
       if (invalidResponse != null && invalidResponse.isFail) {
         _log.warning(
           message: 'TurboWriteable was invalid!',
@@ -109,7 +108,7 @@ extension TurboFirestoreUpdateApi<T> on TurboFirestoreApi<T> {
           message: 'WriteBatch was not null! Updating with batch..',
           sensitiveData: null,
         );
-        final lastBatchResponse = await updateDocs(
+        final lastBatchResponse = await updateDocInBatch(
           writeable: writeable,
           id: id,
           writeBatch: writeBatch,
@@ -126,8 +125,8 @@ extension TurboFirestoreUpdateApi<T> on TurboFirestoreApi<T> {
           message: 'WriteBatch was null! Updating without batch..',
           sensitiveData: null,
         );
-        final documentReference = getDocRefById(
-            id: id, collectionPathOverride: collectionPathOverride);
+        final documentReference =
+            getDocRefById(id: id, collectionPathOverride: collectionPathOverride);
         _log.debug(
           message: 'Creating JSON..',
           sensitiveData: null,
@@ -156,8 +155,7 @@ extension TurboFirestoreUpdateApi<T> on TurboFirestoreApi<T> {
               data: writeableAsJson,
             ),
           );
-          transaction.update(
-              getDocRefById(id: documentReference.id), writeableAsJson);
+          transaction.update(getDocRefById(id: documentReference.id), writeableAsJson);
         }
         _log.info(
           message: 'Updating data done!',
@@ -209,7 +207,7 @@ extension TurboFirestoreUpdateApi<T> on TurboFirestoreApi<T> {
   /// ```dart
   /// final batch = firestore.batch();
   /// final user = User(name: 'John Updated');
-  /// final response = await api.updateDocs(
+  /// final response = await api.updateDocInBatch(
   ///   writeable: user,
   ///   id: 'user-123',
   ///   writeBatch: batch,
@@ -225,9 +223,8 @@ extension TurboFirestoreUpdateApi<T> on TurboFirestoreApi<T> {
   ///
   /// See also:
   /// [updateDoc] single document updates
-  /// [createDocs] batch creation
-  Future<TurboResponse<WriteBatchWithReference<Map<String, dynamic>>>>
-      updateDocs({
+  /// [createDocInBatch] batch creation
+  Future<TurboResponse<WriteBatchWithReference<Map<String, dynamic>>>> updateDocInBatch({
     required TurboWriteable writeable,
     required String id,
     WriteBatch? writeBatch,
@@ -240,8 +237,8 @@ extension TurboFirestoreUpdateApi<T> on TurboFirestoreApi<T> {
       'therefore, you must specify the collectionPathOverride containing all parent collection and document ids '
       'in order to make this method work.',
     );
-    final TurboResponse<WriteBatchWithReference<Map<String, dynamic>>>?
-        invalidResponse = writeable.validate();
+    final TurboResponse<WriteBatchWithReference<Map<String, dynamic>>>? invalidResponse =
+        writeable.validate();
     if (invalidResponse != null && invalidResponse.isFail) {
       _log.warning(
         message: 'TurboWriteable was invalid!',
@@ -281,8 +278,7 @@ extension TurboFirestoreUpdateApi<T> on TurboFirestoreApi<T> {
         writeableAsJson,
       );
       _log.info(
-        message:
-            'Adding update to batch done! Returning WriteBatchWithReference..',
+        message: 'Adding update to batch done! Returning WriteBatchWithReference..',
         sensitiveData: null,
       );
       return TurboResponse.success(